refactor : 미팅 찾기 정리 안된 부분 정리
