+++ src/components/pages/wishlist/List.tsx
@@ -7,7 +7,7 @@ import { useState } from "react"
 import FilterCalendar from "@/components/pages/findMeeting/FilterCalendar/FilterCalendar"
 import FilterSort from "@/components/pages/findMeeting/FilterSort/FilterSort"
 import FilterTab from "@/components/pages/findMeeting/FilterTab/FilterTab"
-import { MeetingCard } from "@/components/pages/findMeeting/MeetingCard/MeetingList/MeetingList"
+import MeetingCard from "@/components/pages/findMeeting/MeetingCard/MeetingCard"
 import ByeBtn from "@/components/pages/wishlist/ByeBtn"
 import Filter from "@/components/public/Filter/Filter"
 import ResetFilter from "@/components/public/ResetFilter"
